set xdebug.file_link_format
This setting determines the format of the links that are made in the
display of stack traces where file names are used. This allows IDEs to
set up a link-protocol that makes it possible to go directly to a line
and file by clicking on the filenames that Xdebug shows in stack traces.